+++ backend/python/pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 description = ""
 authors = ["iamtatsuki05 <tatsukio0522@gmail.com>"]
 packages = [
-    { include = "project", from = "src/" },
+    { include = "grpc_app", from = "src/" },
 ]
 
 [tool.poetry.dependencies]
