+++ backend/python/tests/grpc_app/test_env.py
@@ -1,4 +1,4 @@
-from project.env import VERSION
+from grpc_app.env import VERSION
 
 
 def test_version():
